# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Google Uzbek Voice Agent",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/chats": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "List a device's chats (newest first)",
        "description": "Chat summaries for one device (contract §2.1). A corrupt store\nnever 500s — it degrades to an empty list.",
        "operationId": "list_chats_chats_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Stable per-install device id (UUID) that scopes all chat data.",
              "default": "",
              "title": "User Id"
            },
            "description": "Stable per-install device id (UUID) that scopes all chat data."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Max summaries returned (clamped server-side).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max summaries returned (clamped server-side)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user_id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Create an empty chat",
        "description": "Create a chat for the device and return its summary (contract\n§2.2). The chat is bound to a voice session later via\n``SessionStart.chat_id`` on the WebSocket.",
        "operationId": "create_chat_chats_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user_id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chats/{chat_id}": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Fetch one chat with full message history",
        "description": "Full chat detail: messages, last diagnosis and agronom-review state\n(contract §2.3).",
        "operationId": "get_chat_chats__chat_id__get",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Stable per-install device id (UUID) that scopes all chat data.",
              "default": "",
              "title": "User Id"
            },
            "description": "Stable per-install device id (UUID) that scopes all chat data."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user_id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Unknown chat — or owned by another device (never distinguished).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chats/{chat_id}/agronom-request": {
      "post": {
        "tags": [
          "agronom"
        ],
        "summary": "Farmer requests a human-agronom review",
        "description": "Mark the chat's diagnosis as pending expert review (contract P3.3).\nIdempotent: repeat calls while pending/done change nothing. May start\nthe mock AI second-opinion runner (P3.5) when configured.",
        "operationId": "agronom_request_chats__chat_id__agronom_request_post",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgronomRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user_id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Chat not found, or the agronom feature is disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chats/{chat_id}/agronom-review": {
      "post": {
        "tags": [
          "agronom"
        ],
        "summary": "Expert submits the agronom verdict",
        "description": "Land the human expert's verdict on a pending review (contract\nP3.4) and append the agronom message to the chat. Expert-only —\ngated by ``X-Agronom-Token``.",
        "operationId": "agronom_review_submit_chats__chat_id__agronom_review_post",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          },
          {
            "name": "x-agronom-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expert reviewer token (AGRONOM_REVIEW_TOKEN).",
              "title": "X-Agronom-Token"
            },
            "description": "Expert reviewer token (AGRONOM_REVIEW_TOKEN)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgronomReviewBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user_id / verdict, or empty expert_summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Chat not found, or the agronom feature is disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "409": {
            "description": "The farmer never requested a review for this chat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgronomRequestBody": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Stable per-install device id (UUID) that scopes all chat data.",
            "default": "",
            "examples": [
              "11111111-2222-3333-4444-555555555555"
            ]
          }
        },
        "type": "object",
        "title": "AgronomRequestBody",
        "description": "POST /chats/{chat_id}/agronom-request request body (contract P3.3)."
      },
      "AgronomReviewBody": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Stable per-install device id (UUID) that scopes all chat data.",
            "default": "",
            "examples": [
              "11111111-2222-3333-4444-555555555555"
            ]
          },
          "verdict": {
            "type": "string",
            "title": "Verdict",
            "description": "\"confirmed\" or \"adjusted\".",
            "default": ""
          },
          "expert_summary": {
            "type": "string",
            "title": "Expert Summary",
            "description": "Farmer-facing Uzbek summary, ≤600 chars. Required.",
            "default": ""
          },
          "expert_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expert Notes",
            "description": "Up to 6 notes, ≤300 chars each."
          },
          "adjusted_preparations": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Adjusted Preparations",
            "description": "Frozen P2.1 Preparation shape; [] keeps the AI list."
          }
        },
        "type": "object",
        "title": "AgronomReviewBody",
        "description": "POST /chats/{chat_id}/agronom-review request body (contract P3.4) —\nthe human expert's verdict, submitted with the X-Agronom-Token header."
      },
...
      "ChatDetailResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ChatDetail"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ChatDetailResponse"
      },
      "ChatListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ChatSummary"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ChatListResponse"
      },
      "ChatMessageOut": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "farmer | rais | system | agronom"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "text | question | answer | photo | diagnosis | agronom_review"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url",
            "description": "Public photo URL — only on kind == photo messages."
          }
        },
        "type": "object",
        "required": [
          "role",
          "kind",
          "text",
          "ts"
        ],
        "title": "ChatMessageOut",
        "description": "One stored chat message. ``photo_url`` is present ONLY on photo\nmessages (kind == \"photo\") — omitted otherwise, never null."
      },
      "ChatSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "query_type": {
            "type": "string",
            "title": "Query Type",
            "description": "\"\" | disease_pest | weed | general"
          },
          "crop_id": {
            "type": "string",
            "title": "Crop Id"
          },
          "crop_name": {
            "type": "string",
            "title": "Crop Name"
          },
          "plant_part": {
            "type": "string",
            "title": "Plant Part"
          },
          "symptom_done": {
            "type": "boolean",
            "title": "Symptom Done"
          },
          "symptom_summary": {
            "type": "string",
            "title": "Symptom Summary"
          },
          "general_question": {
            "type": "string",
            "title": "General Question"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count"
          },
          "last_message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatMessageOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "agronom_review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgronomReviewOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "query_type",
          "crop_id",
          "crop_name",
          "plant_part",
          "symptom_done",
          "symptom_summary",
          "general_question",
          "created_at",
          "updated_at",
          "finished",
          "message_count",
          "last_message",
          "agronom_review"
        ],
        "title": "ChatSummary",
        "description": "The list/create response shape (contract §2.1/§2.2)."
      },
      "ChatSummaryResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ChatSummary"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ChatSummaryResponse"
      },
      "CreateChatBody": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Stable per-install device id (UUID) that scopes all chat data.",
            "default": "",
            "examples": [
              "11111111-2222-3333-4444-555555555555"
            ]
          }
        },
        "type": "object",
        "title": "CreateChatBody",
        "description": "POST /chats request body (contract §2.2)."
      },
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "description": "FastAPI's HTTPException shape — all documented error statuses use it."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}